compiles without tests

+++ src/token-voting/LlamaERC20TokenCaster.sol
@@ -31,8 +31,8 @@ contract LlamaERC20TokenCaster is LlamaTokenCaster {
     ERC20Votes _token,
     ILlamaCore _llamaCore,
     uint8 _role,
-    uint256 _voteQuorumPct,
-    uint256 _vetoQuorumPct
+    uint16 _voteQuorumPct,
+    uint16 _vetoQuorumPct
   ) external initializer {
     __initializeLlamaTokenCasterMinimalProxy(_llamaCore, _role, _voteQuorumPct, _vetoQuorumPct);
     token = _token;
