+++ src/token-voting/LlamaTokenVotingFactory.sol
@@ -72,8 +72,8 @@ contract LlamaTokenVotingFactory {
     uint8 actionCreatorRole,
     uint8 casterRole,
     uint256 creationThreshold,
-    uint256 voteQuorumPct,
-    uint256 vetoQuorumPct
+    uint16 voteQuorumPct,
+    uint16 vetoQuorumPct
   ) external returns (address actionCreator, address caster) {
     if (isERC20) {
       actionCreator = address(
@@ -140,8 +140,8 @@ contract LlamaTokenVotingFactory {
     ILlamaCore llamaCore,
     uint256 nonce,
     uint8 role,
-    uint256 voteQuorumPct,
-    uint256 vetoQuorumPct
+    uint16 voteQuorumPct,
+    uint16 vetoQuorumPct
   ) internal returns (LlamaERC20TokenCaster caster) {
     caster = LlamaERC20TokenCaster(
       Clones.cloneDeterministic(
@@ -158,8 +158,8 @@ contract LlamaTokenVotingFactory {
     ILlamaCore llamaCore,
     uint256 nonce,
     uint8 role,
-    uint256 voteQuorumPct,
-    uint256 vetoQuorumPct
+    uint16 voteQuorumPct,
+    uint16 vetoQuorumPct
   ) internal returns (LlamaERC721TokenCaster caster) {
     caster = LlamaERC721TokenCaster(
       Clones.cloneDeterministic(
